Standard error of the log odds ratio, binary predictor

Computes the standard error (and confidence interval) of the log odds ratio for a binary (e.g. treatment) predictor in a proportional-odds model, at a given total sample size n.

se_binary_po(pc, OR, n, r = 1, method = c("ologit", "whitehead"), conf = 0.95)

Arguments

pc

Control-arm category probabilities (sum to 1).

OR

Common odds ratio (experimental vs control arm).

n

Total sample size (both arms).

r

Allocation ratio, treatment:control.

method

"ologit" (default) or "whitehead"

conf

Confidence level.

Value

A list with elements conf, se_log_OR, OR, lower_OR, upper_OR (CI limits for the OR), and method.

Details

Two approaches to the SE are supported, selectable via method:

"whitehead" - Whitehead's (1993) formula

"ologit" - White, Marley-Zagar, Morris, Parmar, Royston & Babiker (2023) fit a weighted proportional-odds model to the anticipated distribution and use the SE from the observed information matrix. This is their "AA" variant.

Examples

pc <- c(0.15, 0.10, 0.10, 0.10, 0.10, 0.45)
se_binary_po(pc, OR = 1.7, n = 30, method = "whitehead")
#> $conf
#> [1] 0.95
#> 
#> $se_log_OR
#> [1] 0.6557662
#> 
#> $OR
#> [1] 1.7
#> 
#> $lower_OR
#> [1] 0.4701752
#> 
#> $upper_OR
#> [1] 6.146645
#> 
#> $method
#> [1] "whitehead"
#> 
se_binary_po(pc, OR = 1.7, n = 30, method = "ologit")
#> $conf
#> [1] 0.95
#> 
#> $se_log_OR
#> [1] 0.6613742
#> 
#> $OR
#> [1] 1.7
#> 
#> $lower_OR
#> [1] 0.4650356
#> 
#> $upper_OR
#> [1] 6.214578
#> 
#> $method
#> [1] "ologit"
#>